Bugs 2620

See merge request cellframe/libdap-chain-global-db!7

+++ dap_chain_global_db_driver.c
@@ -310,7 +310,7 @@ char* dap_chain_global_db_driver_hash(const uint8_t *data, size_t data_size)
     dap_hash_fast(data, data_size, &l_hash);
     size_t a_str_max = (sizeof(l_hash.raw) + 1) * 2 + 2; /* heading 0x */
     char *a_str = DAP_NEW_Z_SIZE(char, a_str_max);
-    size_t hash_len = dap_chain_hash_fast_to_str(&l_hash, a_str, a_str_max);
+    size_t hash_len = (size_t)dap_chain_hash_fast_to_str(&l_hash, a_str, a_str_max);
     if(!hash_len) {
         DAP_DELETE(a_str);
         return NULL;
